gateway/requirements.txt
flask
flask-cors
waitress
requests
flask_sqlalchemy
mariadb